requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.4.5.2
chardet==3.0.4
cssselect==1.1.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
google-api-core==1.22.1
google-api-python-client==1.11.0
google-auth==1.21.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
install==1.3.3
jdcal==1.4.1
lxml==4.5.1
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.5
parse==1.15.0
Pillow==7.1.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
pytz==2020.1
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.46.1
uritemplate==3.0.1
urllib3==1.25.9
w3lib==1.22.0
websockets==8.1
xlrd==1.2.0
XlsxWriter==1.3.3
